rename rust-demangle to rustDemangle to fit naming convention
